removed nextHandlerAction in plan.java, unimported unused imports reverted the changes with TerraformApply.java
