+++ tf-build-tools/tf-cmd-api/pom.xml
@@ -77,7 +77,7 @@
                     <limit implementation="org.jacoco.report.check.Limit">
                       <counter>LINE</counter>
                       <value>COVEREDRATIO</value>
-                      <minimum>0.50</minimum>
+                      <minimum>0.80</minimum>
                     </limit>
                   </limits>
                 </rule>
